% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohenKappa.R
\name{CohenKappa}
\alias{CohenKappa}
\title{Calculate Cohen's Kappa.}
\usage{
CohenKappa(lambda, piA1, piI1)
}
\arguments{
\item{lambda}{the proportion of truly activated voxels}

\item{piI1}{the probability that a truly inactive voxel is declared active}

\item{piAI}{the probability that a truly active voxel is declared active}
}
\value{
Numeric with the index of coherence (\eqn{\kappa}).
}
\description{
Function to calculate a coherence incex called Cohen's \eqn{\kappa} which
gives an estimate of the amount of agreement between different raters.
Or the agreement on realizations of observations from a mixture of two
binomial distributions.
}
\details{
This function is used to estimate the agreement of replication images in fMRI data.
Therefore, we denote the parameters that are obtained and described in \code{\link{EMbinom}}.
}
