% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrJ.R
\name{corrJ}
\alias{corrJ}
\title{Correction factor J}
\usage{
corrJ(N)
}
\arguments{
\item{N}{sample size in this study}
}
\value{
A positive integer J
}
\description{
Helper function for \code{\link{hedgeG}}.
Calculates the correction factor in the conversion from a t-statistic
to g.
}
