% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupLevelOLS.R
\name{GroupLevelOLS}
\alias{GroupLevelOLS}
\title{Perform group level model using OLS}
\usage{
GroupLevelOLS(N, subj_b1)
}
\arguments{
\item{N}{an integer specifying the amount of subjects from the first level}

\item{subj_b1}{a vector of length N with the estimated beta 1 parameters}

\item{...}{extra arguments passed to or from other methods.}
}
\value{
a vector with the estimated group level parameter
}
\description{
This function performs simple OLS on the beta 1 parameters obtained fitting a GLM at subject level.
The user does not need to specify a design, as this function is only used
to calculate one sample tests (corresponding to t-tests).
}
