% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hedgeG.R
\name{hedgeG}
\alias{hedgeG}
\title{Calculate Hedges' g from one sample t-test statistic.}
\usage{
hedgeG(t, N, type = c("approx", "exact"))
}
\arguments{
\item{t}{T test-statistic from a one-sample t-test}

\item{N}{Sample size.}

\item{type}{Use the approximation for the unbiased estimator (default)
or the exact expression.}
}
\value{
Standardized mean effect size Hedges' g.
}
\description{
This function takes the t test-statistic from a \strong{one-sample t-test}
and calculates Hedges' g.
}
\details{
The approximation is based on the correction factor \code{\link{corrJ}}.
The exact expression is based on \code{\link{corrH}}.
}
