% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlapAct.R
\name{overlapAct}
\alias{overlapAct}
\title{Calculate overlap of activation (Maitra's adaptation of Dice).}
\usage{
overlapAct(bin_map1, bin_map2)
}
\arguments{
\item{bin_map1}{the first binarized (thresholded) map}

\item{bin_map2}{the second binarized (thresholded) map}
}
\value{
Numeric value: measure of overlap.
}
\description{
Function to calculate the overlap of activation of two binarized images.
The formula is an adaptation of the Dice overlap, suggested by Maitra in 2010.
}
\details{
The measure is the number of hits (e.g. voxels) intersecting in both images
divided by the sum of the number of hits in each image minus its intersection again:
\deqn{V_{a,b}/(V_a + V_b - V_{a,b})}
Reference:
Maitra, R. (2010). A re-defined and generalized percent-overlap-of- activation measure for studies of fMRI reproducibility and its use in identifying outlier activation maps. Neuroimage 50, 124–135.
}
