% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_keywords_GS.R
\name{plotKeyWord}
\alias{plotKeyWord}
\title{Plot trend of keywords in Google Scholar.}
\usage{
plotKeyWord(keyword, start, end)
}
\arguments{
\item{keyword}{the keyword you would like to search}

\item{start}{startdate (in year, e.g. 2000)}

\item{end}{enddate (in year, e.g. 2017)}
}
\description{
This function does something.
}
