% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimPercBOLD.R
\name{SimPercBOLD}
\alias{SimPercBOLD}
\title{Simulate an fMRI time series using a parametric GLM with known beta parameters.}
\usage{
SimPercBOLD(percBOLD, designRange, nscan, sigmaNoise)
}
\arguments{
\item{percBOLD}{a number (in percentage) providing the true percent change of BOLD change.}

\item{designRange}{a vector with the min and max of the design.}

\item{nscan}{a number providing the number of scans}

\item{sigmaNoise}{a single number representing sigma of white noise}

\item{...}{extra arguments passed to or from other methods.}
}
\value{
a list with a vector with the simulated time series, the signal, the design matrix
}
\description{
This temporarily function creates a time series of nscan for one subject.
The user can specify the % change of BOLD signal as beta 1.
Then it specifies the range of the design matrix and sigma for white noise.
It can be used to run quick simulations.
}
