% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau.R
\name{tau}
\alias{tau}
\title{DerSimonian & Laird estimator for between-study variance.}
\usage{
tau(Y, W, k)
}
\arguments{
\item{Y}{a vector with the estimates for the study effects
(e.g. standardized effect sizes Hedges' g).}

\item{W}{a vector with the study specific weights.
Most often the inverse of the variance of Y.
See also \code{\link{varHedge}}.}

\item{k}{amount of studies in the meta-analysis.}
}
\value{
an estimate for between-study variance.
}
\description{
Function to estimate \eqn{\tau^2}, the DerSimonian & Laird (1986) estimator
for between-study variance. Usefull to execute a random effects meta-analysis.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pubmed/3802833}
}
