% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varHedge.R
\name{varHedge}
\alias{varHedge}
\title{Calculate within-study variance of g}
\usage{
varHedge(g, N)
}
\arguments{
\item{g}{standardized effect size g from one sample set-up.}

\item{N}{sample size}
}
\value{
The estimate of the within-study variability of Hedges g.
}
\description{
Function to calculate the within-study variance of one sample
effect size g using the formula from Radua.
Input is the standardized effect size g and N.
}
\seealso{
\code{\link{varHedgeT}} for the function of the variance based on
the non-central t-distribution from which we know the derivations.
We do not know the derivations for the current expression from Radua.
}
