% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunUnivMA.R
\name{wMean}
\alias{wMean}
\title{Calculate a weighted average}
\usage{
wMean(Y, W)
}
\arguments{
\item{Y}{the values to calculate the weighted average over.}

\item{W}{the weights (of length Y or 1) to use in the weighted average}
}
\value{
the weighed average
}
\description{
Function to calculate a weighted average.
}
