% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helperFunUnivMA.R
\name{hedgeG}
\alias{hedgeG}
\title{Calculate Hedges' g from one sample t-test statistic.}
\usage{
hedgeG(t, N)
}
\arguments{
\item{t}{t test-statistic from a one-sample t-test}

\item{N}{sample size.}
}
\value{
Hedges' g.
}
\description{
This function takes the t test-statistic from a \strong{one-sample t-test}
and calculates Hedges' g.
}
\seealso{
\code{\link{corrJ}} for the function of the correction factor J.
}
