% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varHedge.R
\name{varHedge}
\alias{varHedge}
\title{Calculate within-study variance of g}
\usage{
varHedge(g, N)
}
\arguments{
\item{g}{standardized effect size g from one sample set-up.}

\item{N}{sample size}
}
\value{
var(g)
}
\description{
Function to calculate the within-study variance of one sample
effect size g using calculations from Hedges (1989).
Input is the standardized effect size g and N.
}
